requirements.txt
haystack-ai
mcp-haystack
python-dotenv